separated statespace init & verification
this allows for the definition of models whose statespace may not
satisfy the Blanchard-Kahn rank or order conditions, which is desirable
in some cases; most notably (at this point), the solutions of models
with occasionally binding constraints. this also allows for the
jacobian & 'hessian' (misnomer) of a model to be used, too.