updated implementation of VarArray
previously, the VarArray class had a 'variables' attribute that stored
the instances of ModelVar's. this attribute was referenced in the
__array_function__ class method when concatenating VarArray's, throwing
an AttributeError. this has been fixed.
    although DetermVars can be parsed, not much has been done with them.
hence, when accessing a Model's 'exog' parameter, the old method of
concatenating two lists of variable names was used. now, the VarArrays
are properly concatenated